fix: resolve iOS camera lifecycle issues and add Android remote image support
  iOS Camera Fixes:
  - Fix camera initialization to respect cameraFacing prop on first load
  - Prevent main thread blocking by moving AVCaptureSession start/stop to background queue
  - Add camera session state tracking (isRunning, isDeallocating, hasStarted)
  - Add guards to prevent starting already-running camera sessions
  - Fix camera position switching causing blank feed
  - Improve deinit cleanup to prevent crashes during rapid open/close cycles
  - Remove excessive debug logging

  Android Improvements:
  - Add HTTP/HTTPS support for remote image URLs in uriToBitmap()
  - Update onImageSharpnessScore signature to Double for SDK compatibility

  Native SDK Updates:
  - Update Android VisionSDK to v 2.4.25
  - Update iOS VisionSDK to 2.0.2

+++ android/src/main/java/com/visionsdk/VisionCameraViewManager.kt
@@ -495,7 +495,7 @@ class VisionCameraViewManager(private val appContext: ReactApplicationContext) :
       // Not used in minimal implementation
     }
 
-    override fun onImageSharpnessScore(imageSharpnessScore: Float) {
+    override fun onImageSharpnessScore(imageSharpnessScore: Double) {
       Log.d(TAG, "onImageSharpnessScore called: $imageSharpnessScore")
       val event = Arguments.createMap()
       event.putDouble("sharpnessScore", imageSharpnessScore.toDouble())
